+++ OpenCVSharpPractise/OpenCVSharpPractise/ImageProccessingApp.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34723.18
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ImageProccessingApp", "ImageProccessingApp\ImageProccessingApp.csproj", "{CF106F4A-9DE7-4B5D-9411-A3A9E773A486}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CF106F4A-9DE7-4B5D-9411-A3A9E773A486}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF106F4A-9DE7-4B5D-9411-A3A9E773A486}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF106F4A-9DE7-4B5D-9411-A3A9E773A486}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF106F4A-9DE7-4B5D-9411-A3A9E773A486}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1E0417D0-AB73-4F07-AC1D-7E3B44C7858D}
+	EndGlobalSection
+EndGlobal
