+++ OpenCVSharpPractise/OpenCVSharpPractise/ImageProccessingApp/Program.cs
@@ -1,0 +1,17 @@
+namespace ImageProccessingApp
+{
+    internal static class Program
+    {
+        /// <summary>
+        ///  The main entry point for the application.
+        /// </summary>
+        [STAThread]
+        static void Main()
+        {
+            // To customize application configuration such as set high DPI settings or default font,
+            // see https://aka.ms/applicationconfiguration.
+            ApplicationConfiguration.Initialize();
+            Application.Run(new Form1());
+        }
+    }
+}
